feat: implements input file parsing
Given the path to a Q-Chem input file, the QchemInput class will
parse and store the contents as a dict with $sections as the keys.
The values are dics containing as their keys the first keyword on
each line and as their values, a list of the remainder of the line.

Basic parsing of the $XC_FUNCTIONAL section creates a dict of the
following structure:
{
	"X": {
		"FUNCTIONAL": float component,
		"FUNCTIONAL": float component,
		...
	},
	"C": { ... }
}