feat: adds functioning setup and submission of tuning calculation
The dict structure of the input has been transitioned to a list,
since certain input sections will contain duplicate keywords.

The QchemCalculation class sets up and optionally runs a Qchem
calculation, given an input file. Note that in order for this
to function, the user MUST have the Qchem binaries on their PATH.
In other words, the user must be able to run a Qchem calculation
by invoking:
	qchem -save 1 -nt THREADS file.in file.out jobname

The QchemTuning class takes a given NEUTRAL input file and value
for the range separation parameter, omega, and creates three files,
one each for neutral, aniona and cation, at that value of omega.
The multiplicities for each species can be specified, otherwise
a neutral singlet and anion/cation doublets are assumed. The class
then allows for calculations for each species to be run sequentially.