feat: adds fully fledged tuning and testing capabilities
The driver.py file is intended to serve as the interface for most
users and allows for automated single point and range calculations
as well as tabulation of prior results.

The test/test.py file implements tests of the major classes within
the RSHtune package and demonstrates their functionality. Along
with this file, an example input and geometry are provided.

The Makefile simplifies addition of the package to the PYTHONPATH
to facilitate location independent import.